Added texture support and model based rendering

+++ Shmungus/src/loadingTools/Vertex.h
@@ -41,7 +41,7 @@ struct ColorQuadVertex {
 struct EntityVertex {
 
 	float positions[3];
-	byte texCoords[2];
-	byte textureID;
+	float texCoords[2];
+	float textureID;
 
 };
